+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 FEATURES:
 
+  - Puppet provisioner now supports Hiera by specifying a `hiera_config_path`.
   - Added a `working_directory` configuration option to the Puppet apply
     provisioner so you can specify the working directory when `puppet` is
     called, making it friendly to Hiera data and such. [GH-1670]